refactor: 콘솔 패키지 util로 변경
